Add support for sourcing repository credentials from the keyring.
This commit updates both the metadata and package caches to support
downloading packages and package indexes from repositories that require
HTTP basic authentication to access.

Initial support for these authenticated repositories is very narrow: the
repository URL must contain a username, no password, and have an entry
in the system keyring. We also don't make any attempt to prompt users
for credentials when requests fail.

Unit tests are included for the new authentication header helpers, but
there are currently no tests of end-to-end workflows with an
authenticated repository, and I may have missed something.

Part of r-lib/pak#729.

Signed-off-by: Aaron Jacobs <aaron.jacobs@posit.co>